% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_rows.R
\name{dup_rows}
\alias{dup_rows}
\title{A row duplicating function to weight data (for statistics and maps)}
\usage{
dup_rows(df, col)
}
\arguments{
\item{df}{a data frame containing metadata (presumably for zebras and cattle) and containing the column "Total.zebras"}
}
\value{
a new data frame with duplicated rows
}
\description{
A row duplicating function to weight data (for statistics and maps)
}
\examples{
  dup_rows(ws)
}
