% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wb.R
\name{generate_wb}
\alias{generate_wb}
\title{A function to generate the Wildbook metadata spreadsheet}
\usage{
generate_wb(white, PATH, date)
}
\arguments{
\item{white}{a data frame containing metadata for zebras that can be used to fill out a Wildbook spreadsheet}

\item{PATH}{the path name telling you where the photos are (i.e., "/Users/annikahsi/Desktop/SUMMER 2022/photos/28.6.22N/group")}

\item{date}{the date in the format "dd/m/yyyy"}
}
\value{
the filled in wildbook data frame
}
\description{
A function to generate the Wildbook metadata spreadsheet
}
