% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_ws_match.R
\name{vs_ws_match}
\alias{vs_ws_match}
\title{A function to match vegetatino to the nearest animal sighting on the same day}
\usage{
vs_ws_match(ws, vs)
}
\arguments{
\item{ws}{a data frame containing metadata (presumably on zebras and cattle) containing rows "Latitude" and "Longitude"}

\item{vs}{a data frame containing statistics (presumably on vegetation) containing rows "Latitude" and "Longitude"}
}
\value{
data frame containing vs rows with corresponding ws rows
}
\description{
A function to match vegetatino to the nearest animal sighting on the same day
}
