% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_dist.R
\name{water_dist}
\alias{water_dist}
\title{A function to calculate the distance between given coordinates and water locations at Mpala}
\usage{
water_dist(df, water_loc)
}
\arguments{
\item{df}{a data frame containing metadata (presumably for zebras and cattle) and containing the columns "Date," "Latitude," "Longitude," and "Distance.to.water"}

\item{water_loc}{a data frame of coordinates (Latitude/Longitude) for water locations (presumably in Mpala)}
}
\value{
a data frame identical to df but with the Distance.to.water column filled in
}
\description{
A function to calculate the distance between given coordinates and water locations at Mpala
}
